requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
command-not-found==0.3
dbus-python==1.2.18
discord.py==2.4.0
distro==1.7.0
distro-info===1.1build1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
netifaces==0.11.0
PyGObject==3.42.1
python-apt==2.4.0+ubuntu1
python-dotenv==1.0.1
PyYAML==5.4.1
requests==2.32.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
